requirements.txt
numpy
matplotlib
pandas
librosa
mir_eval
IPython
pygame-ce